# manage-service/docs/swagger.yaml
definitions:
...
  entity.Document:
    properties:
      applicant_id:
        type: integer
      file_name:
        type: string
      file_type:
        type: string
      id:
        type: integer
      status:
        type: string
      storage_path:
        type: string
      uploaded_at:
        type: string
    type: object
  entity.DocumentQueueTask:
    properties:
      applicant_id:
        type: integer
      created_at:
        type: string
      document_category:
        type: string
      error_message:
        type: string
      file_path:
        type: string
      id:
        type: string
      priority:
        type: integer
      status:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API for managing applicants and documents
  title: Manage Service API
  version: "1.0"
paths:
...
  /v1/applicants/{id}/documents:
    get:
      description: Возвращает список всех загруженных документов абитуриента с их
        метаданными и статусами обработки.
      parameters:
      - description: ID абитуриента
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.Document'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Список документов абитуриента
      tags:
      - documents
    post:
      consumes:
      - multipart/form-data
      description: Загружает файл документа для абитуриента, сохраняет в MinIO и запускает
        очередь AI-обработки.
      parameters:
      - description: ID абитуриента
        in: path
        name: id
        required: true
        type: integer
      - description: Категория документа
        in: formData
        name: category
        required: true
        type: string
      - description: Тип документа (например, для профессиональной переподготовки)
        in: formData
        name: doc_type
        type: string
      - description: Файл документа (PDF)
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Загрузка документа
      tags:
      - documents
  /v1/applicants/{id}/documents/{docId}:
    delete:
      description: Удаляет документ и все связанные с ним извлечённые данные по ID
        абитуриента и ID документа.
      parameters:
      - description: ID абитуриента
        in: path
        name: id
        required: true
        type: integer
      - description: ID документа
        in: path
        name: docId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Удаление документа
      tags:
      - documents
  /v1/applicants/{id}/documents/reprocess:
    post:
      description: Запускает повторное AI-извлечение для последнего загруженного документа
        абитуриента по указанной категории.
      parameters:
      - description: ID абитуриента
        in: path
        name: id
        required: true
        type: integer
      - description: Категория документа
        in: query
        name: category
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Повторная обработка последнего документа
      tags:
      - documents
  /v1/applicants/{id}/documents/view:
    get:
      description: Возвращает содержимое последнего загруженного файла по ID абитуриента
        и категории. Content-Type определяется типом файла.
      parameters:
      - description: ID абитуриента
        in: path
        name: id
        required: true
        type: integer
      - description: Категория документа
        in: query
        name: category
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Просмотр последнего документа абитуриента
      tags:
      - documents
...
  /v1/applicants/{id}/queue-status:
    get:
      description: Возвращает список задач в очереди AI-обработки документов для указанного
        абитуриента.
      parameters:
      - description: ID абитуриента
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.DocumentQueueTask'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Статус очереди обработки
      tags:
      - documents
...
  /v1/documents/{id}/category:
    patch:
      description: Меняет категорию документа на новую и запускает его повторную обработку.
      parameters:
      - description: ID документа
        in: path
        name: id
        required: true
        type: integer
      - description: 'Новая категория: { \'
        in: body
        name: category
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Изменение категории документа
      tags:
      - documents
  /v1/documents/{id}/reprocess:
    post:
      description: Запускает повторное AI-извлечение для конкретного документа по
        его ID.
      parameters:
      - description: ID документа
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Повторная обработка документа по ID
      tags:
      - documents
  /v1/documents/{id}/status:
    get:
      description: Возвращает текущий статус обработки документа по его ID.
      parameters:
      - description: ID документа
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Статус документа
      tags:
      - documents
  /v1/documents/{id}/view:
    get:
      description: Возвращает содержимое файла документа напрямую (inline) по ID документа.
        Content-Type определяется типом файла.
      parameters:
      - description: ID документа
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Просмотр документа по ID
      tags:
      - documents
...
securityDefinitions:
  BearerAuth:
    description: 'Введите токен в формате: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"